+++ sdk/appconfiguration/azure-appconfiguration/CHANGELOG.md
@@ -11,7 +11,7 @@
 - Fixed async `update_sync_token` to use async/await keywords
 
 ### Other Changes
-- Set the default `api_version` to "1.0".
+- Updated the default `api_version` to "2022-11-01-preview".
 - Bumped minimum dependency on `azure-core` to `>=1.25.0`.
 
 ## 1.4.0 (2022-02-13)